+++ packages/shared/components/Sidebar.svelte
@@ -10,7 +10,7 @@
         NotificationBadge,
     } from 'shared/components'
     import { appVersionDetails, mobile } from '@core/app'
-    import { getInitials, isRecentDate } from '@lib/helpers'
+    import { getInitials, isRecentDate } from '@core/utils'
     import { activeProfile } from '@core/profile'
     import {
         dashboardRouter,
