+++ packages/shared/components/atoms/tiles/tileContents/FoundryActivityTileContent.svelte
@@ -1,7 +1,7 @@
 <script lang="typescript">
     import { localize } from '@core/i18n'
     import { InclusionState, IPersistedAsset } from '@core/wallet'
-    import { truncateString } from '@lib/helpers'
+    import { truncateString } from '@core/utils'
     import { Text, AssetIcon, FontWeight } from 'shared/components'
 
     export let amount: string
