+++ packages/shared/components/atoms/tiles/tileContents/TransactionActivityTileContent.svelte
@@ -2,7 +2,7 @@
     import { localize } from '@core/i18n'
     import { networkHrp } from '@core/network'
     import { ActivityDirection, InclusionState, IPersistedAsset, ITransactionActivityData, Subject } from '@core/wallet'
-    import { truncateString } from '@lib/helpers'
+    import { truncateString } from '@core/utils'
     import { Text, AssetIcon, FontWeight } from 'shared/components'
 
     export let amount: string
