+++ packages/shared/components/organisms/NodeConfigurationForm.svelte
@@ -4,7 +4,7 @@
     import { INode, checkNodeUrlValidity, checkNetworkId, IClientOptions, nodeInfo } from '@core/network'
     import { localize } from '@core/i18n'
     import { getNodeInfo } from '@core/profile-manager'
-    import { stripSpaces, stripTrailingSlash } from '@lib/helpers'
+    import { stripSpaces, stripTrailingSlash } from '@core/utils'
     import { activeProfile } from '@core/profile'
 
     export let node: INode = { url: '', auth: { username: '', password: '', jwt: '' } }
