+++ packages/shared/components/popups/ActivityDetailsPopup.svelte
@@ -17,7 +17,7 @@
     import { currencies, exchangeRates } from '@lib/currency'
     import { CurrencyTypes } from 'shared/lib/typings/currency'
     import { setClipboard } from '@core/utils'
-    import { truncateString } from '@lib/helpers'
+    import { truncateString } from '@core/utils'
     import { closePopup, openPopup } from '@auxiliary/popup'
     import { onMount } from 'svelte'
