+++ packages/shared/components/popups/CreateAccountPopup.svelte
@@ -1,6 +1,6 @@
 <script lang="typescript">
     import { Button, ColorPicker, Input, Text } from 'shared/components'
-    import { getTrimmedLength } from 'shared/lib/helpers'
+    import { getTrimmedLength } from '@core/utils'
     import { localize } from '@core/i18n'
     import { closePopup, updatePopupProps } from '@auxiliary/popup'
     import { checkActiveProfileAuth } from '@core/profile'
