+++ packages/shared/lib/core/account/utils/validateAccountName.ts
@@ -1,6 +1,6 @@
 import { localize } from '@core/i18n'
 import { activeAccounts } from '@core/profile'
-import { getTrimmedLength } from '@lib/helpers'
+import { getTrimmedLength } from '@core/utils'
 import { get } from 'svelte/store'
 import { MAX_ACCOUNT_NAME_LENGTH } from '../constants'
