+++ packages/shared/lib/core/app/stores/has-completed-app-setup.store.ts
@@ -1,3 +1,3 @@
-import { persistent } from '@lib/helpers'
+import { persistent } from '@core/utils/store'
 
 export const hasCompletedAppSetup = persistent<boolean>('hasCompletedAppSetup', false)
