+++ packages/shared/lib/core/app/stores/terms-of-service.store.ts
@@ -1,4 +1,4 @@
-import { persistent } from '@lib/helpers'
+import { persistent } from '@core/utils/store'
 
 /**
  * The store containing the last accepted version of the Terms of Service.
