+++ packages/shared/lib/core/profile/stores/active-profile-id.store.ts
@@ -1,3 +1,3 @@
-import { persistent } from '@lib/helpers'
+import { persistent } from '@core/utils/store'
 
 export const activeProfileId = persistent<string>('activeProfileId', null)
