+++ packages/shared/lib/core/profile/stores/profiles.store.ts
@@ -1,4 +1,5 @@
-import { persistent } from '@lib/helpers'
+import { persistent } from '@core/utils/store'
+
 import { IPersistedProfile } from '../interfaces'
 
 export const profiles = persistent<IPersistedProfile[]>('profiles', [])
