+++ packages/shared/lib/core/utils/dom.ts
@@ -1,0 +1,6 @@
+/**
+ * Update application path
+ */
+export function goto(path: string): void {
+    window.location.hash = path
+}
