+++ packages/shared/lib/core/utils/index.ts
@@ -6,6 +6,7 @@ export * from './types'
 export * from './convert'
 export * from './crypto'
 export * from './date'
+export * from './dom'
 export * from './encode'
 export * from './filter'
 export * from './format'
@@ -17,6 +18,7 @@ export * from './object'
 export * from './os'
 export * from './random'
 export * from './sort'
+export * from './store'
 export * from './string'
 export * from './time'
 export * from './ui'
