+++ packages/shared/lib/core/utils/interfaces/date-difference.interface.ts
@@ -1,4 +1,4 @@
-export interface DateDifference {
+export interface IDateDifference {
     unit: string
     value?: number
 }
