+++ packages/shared/lib/core/utils/interfaces/index.ts
@@ -1,2 +1,3 @@
+export * from './date-difference.interface'
 export * from './dropdown.interface'
 export * from './element.interface'
