+++ packages/shared/lib/core/utils/validation.ts
@@ -14,3 +14,27 @@ export function isValidUrl(url: string): boolean {
 export function isValidHttpsUrl(url: string): boolean {
     return validUrl.isHttpsUri(url)
 }
+
+/**
+ * Does the string contain invalid filename chars
+ * @param name The name to validate
+ * @returns
+ */
+export function validateFilenameChars(name: string | undefined): string {
+    if (!name) {
+        return 'emptyName'
+    }
+    if (name.startsWith('~')) {
+        return 'tilde'
+    }
+    /* eslint-disable no-control-regex */
+    if (/[\u0000-\u001f\u0080-\u009f]/g.test(name)) {
+        return 'control'
+    }
+    if (/^\.\./.test(name)) {
+        return 'startDot'
+    }
+    if (/[<>:"/\\|?*]/g.test(name)) {
+        return 'chars'
+    }
+}
