+++ packages/shared/lib/core/wallet/stores/hidden-activities.store.ts
@@ -1,6 +1,8 @@
-import { activeProfileId } from '@core/profile'
-import { persistent } from '@lib/helpers'
 import { get } from 'svelte/store'
+
+import { activeProfileId } from '@core/profile'
+import { persistent } from '@core/utils/store'
+
 import type { IHiddenActivities } from '../interfaces'
 
 export const hiddenActivities = persistent<IHiddenActivities>('hiddenActivities', {})
