+++ packages/shared/lib/core/wallet/utils/getAssetInitials.ts
@@ -1,6 +1,6 @@
 import { IPersistedAsset } from '@core/wallet'
 import { MAX_ASSET_DISPLAYED_INITIALS } from '@core/wallet/constants'
-import { getInitials } from '@lib/helpers'
+import { getInitials } from '@core/utils'
 
 export function getAssetInitials(asset: IPersistedAsset): string {
     return (
