+++ packages/shared/lib/typings/index.ts
@@ -1,6 +1,5 @@
 import * as currency from './currency'
 import * as events from './events'
 import * as message from './message'
-import * as wallet from './dateDifference'
 
-export { currency, events, message, wallet }
+export { currency, events, message }
